// MyMoviePlan-Frontend/mymovieplan/src/app/Models/TempSelectMembers.model.ts
export interface TempSelectMembers {
    auditoriumId: number;
    auditoriumName: string;
    showId: number,
    showName: string;
    showTime: string;
    movieShowId: number;
    bookedSeats: number;
    bookedSeatNumbers: string[];
    date: Date;
    movieName: string;
    movieId: number;
    movieLanguage: string;
    seats: number;
}